pyproject.toml
[tool.poetry]
name = "kgtorrent"
version = "0.1.0"
description = ""
authors = ["Juan Carlos Sigler Priego <juan@repl.it>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
sqlalchemy = "^1.3.20"
sqlalchemy-utils = "^0.40.0"
pymysql = "^1.0.2"
tqdm = "^4.65.0"
requests = "^2.28.2"
kaggle = "^1.5.13"
cryptography = "^40.0.0"
google-cloud-storage = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
